cassandra-load/requirements.txt
cassandra-driver>=3.25
Faker>=15.3